update function to load on document content loaded
