+++ package.json
@@ -27,5 +27,8 @@
     "webpack": "^5.72.1",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.9.0"
+  },
+  "dependencies": {
+    "uniqid": "^5.4.0"
   }
 }
