+++ src/index.js
@@ -1,44 +1,34 @@
-
+import todos from './todoItems';
 import './style.css';
 
 
-const todos = [
-  {
-    description: 'Creating a frontend project',
-    completed: false,
-    index: 1,
-  },
-  {
-    description: 'Baking cakes',
-    completed: false,
-    index: 2,
-  },
-  {
-    description: 'Reading about local storage',
-    completed: false,
-    index: 3,
-  },
-  {
-    description: 'Cleaning the house',
-    completed: false,
-    index: 4,
-  },
-];
 
-todos.forEach((todo) => {
+
+const todoItem = ({description, index}) => {
   const liEl = document.createElement('li');
   liEl.classList.add('todo-wrapper');
+  liEl.id = `todo-${index}`
   liEl.innerHTML = `
+    <div class='left-align'>
       <div>
-        <input type="checkbox" id="${todo.index}" />
+        <input type="checkbox" id="${index}" />
       </div>
       <div>
-        <p contenteditable="true">${todo.description}</p>
+        <p contenteditable="true">${description}</p>
       </div>
+    </div>
       <div>
         <ion-icon name="ellipsis-vertical-outline"></ion-icon>
     </div>
       `;
-      const listContainer = document.querySelector('.todo-list-holder');
-      listContainer.appendChild(liEl);
-});
+    return liEl;
+}
+
+const displayTodo = () => {
+  const listContainer = document.querySelector('.todo-list-holder');
+  todos.forEach((todo) => {
+  listContainer.appendChild(todoItem(todo));
+  });
+}
+
+document.addEventListener('DOMContentLoaded', displayTodo)
