+++ src/todoItems.js
@@ -1,0 +1,25 @@
+
+const todos = [
+    {
+      description: 'Creating a handheld batton',
+      completed: false,
+      index: 1,
+    },
+    {
+      description: 'Baking cakes',
+      completed: false,
+      index: 2,
+    },
+    {
+      description: 'Reading about local storage',
+      completed: false,
+      index: 3,
+    },
+    {
+      description: 'Cleaning the house',
+      completed: false,
+      index: 4,
+    },
+  ];
+
+  export default todos;
